Export works for GIF, PNG, JPG, JPEG
Tried to make it save as PNG by default but it doesn't like it for some
reason. I'll look at it more later, too tired tonight. As long as you
write the extension in the file it works great. Only has an issue if you
don't type the extension at the end.