Grayscale filter works properly; Undo/Redo bug
Set up a grayscale filter in the Filter Menu that properly applies the
filter to all of the previous History objects in the Current Layer. We
can eventually extend the functionality so that it applies to specific
History objects, or even to a selection of the current drawing.

Updated README, Undo/Redo is bugged.