Fixed Export Bug
The default save type is PNG. If no type is written by the user it is
saved to the correct filetype. If the file type given by the user
differs from the one selected, it correctly saves to the user-specified
file type (although we could change this so if you selected PNG image
and saved as "test.jpg" that it would save as "test.jpg.png" and still
be a PNG image.