Cleaned up filters in History class.
Problem to be addressed: in the applyFilter method line 214 in History.java.
Right now if we apply the filter to the selected history object's action
images then we don't change them so if I apply the negative filter more
than once for example, it continues to apply it to the unchanged history
objects so you continue to get the same thing. I commented out those lines
and instead have a line that always applies the filter to the last snapshot,
which allows for us to apply multiple filters properly but removes the
ability to apply to a selection of history objects.