Added Checkered Background to Drawing Area.
Originally tried to do a pattern with octagons but it didn't look good
at all. We can leave it like this for now and update it to look better
and/or more original later. If either of you have a better pattern
suggestion I'd love to try it, but at the moment I don't have a better
idea than this.