Crop tool, and IO updates.
Crop tool currently works on any selected layers (does not work on selected
History objects, and shouldn't). For now it just crops to the size of the
DrawingArea and crops off any "hidden" information. It is set up to support
cropping a selected area once we write our Select Tool.

IO update: When reading a file, it always draws it to a BufferedImage of
TYPE_INT_ARGB. Before when importing a JPG/JPEG, no alpha channel existed
which caused our negative filter to break.